refactor(api): move reconcileSupOrganizationLearner usecase into prescription
